requirements.txt
streamlit~=1.56.0
curl-cffi~=0.15.0
pillow~=12.2.0